Complete dbHelpers and Utils. Utils are the functions that will be used by the server. Helpers are small and mostly useless on their own
